Moved get_archive() out of odt.rs
This is to allow other parsers to reuse it (docx for example)
