+++ server/src/parsers/mod.rs
@@ -1,1 +1,3 @@
 pub mod odt;
+
+mod util;
